Move fixtures for api tests to conftest
